requirements.txt
# ==============================================
# ABoroOffice Requirements - Platform Selection
# Django 6.0.1 (Latest) as base for all integrated applications
# ==============================================
#
# IMPORTANT: Use platform-specific requirements files:
#
#   For Windows 11 Development:
#     pip install -r requirements-windows.txt
#
#   For Linux Production/Deployment:
#     pip install -r requirements-linux.txt
#
# This file serves as a reference for all dependencies.
# Platform-specific versions have been created to handle:
#   - Windows compatibility (pydantic 1.10.13 with pre-built wheels)
#   - Linux optimization (pydantic 2.5.3 with Rust compilation)
#   - OS-specific packages (e.g., uwsgi for Linux only)
#
# ============== REFERENCE: Complete Dependency List ==============

# Django Core - Latest Version
Django==6.0.1
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.11  # PostgreSQL (recommended)
# mysqlclient==2.2.7  # MySQL/MariaDB (optional)
# mssql-django==1.4.2  # MSSQL (optional, requires ODBC driver)

# REST Framework - Django 6.0 Compatible
djangorestframework==3.14.0
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.4.0

# Admin Enhancements - Django 6.0 Compatible
django-jazzmin==3.0.1
django-admin-interface==0.28.0

# Cache & Celery - Django 6.0 Compatible
django-redis==5.4.0
redis==7.1.0
celery==5.6.2
# django-celery-beat==2.8.1  # ⚠️ NOT COMPATIBLE WITH DJANGO 6.0 - Requires Django<6.0. Waiting for maintainer update.
django-celery-results==2.5.1

# WebSockets & Async - Django 6.0 Compatible
channels==4.3.2
channels-redis==4.2.1
daphne==4.1.0

# Authentication & Permissions - Django 6.0 Compatible
django-guardian==3.2.0
django-oauth-toolkit==2.3.0
django-allauth==0.62.0

# HelpDesk Dependencies
django-ckeditor==6.7.3
anthropic==0.30.1
# ollama==0.6.1  # ⚠️ NOT in Windows requirements (pydantic conflict)
                # Linux only: Works with pydantic 2.5.3
                # Windows: Run as standalone service via HTTP API
django-filter==24.1
python-dateutil==2.8.2

# AI & LLM Providers
# Anthropic Claude (via anthropic library above)
# AWS Bedrock (synchronous client - standard usage)
boto3==1.34.18
# aioboto3==14.0.0  # ⚠️ Optional: Only for async AWS operations (version conflict with boto3)
                    # Use standard boto3.client() for most cases
                    # Async support can be added later with compatible versions

# Cloude Dependencies
Pillow==10.4.0
django-imagekit==4.1.0

# Approvals Dependencies
paramiko==4.0.0
pandas==2.3.3
openpyxl==3.1.5
python-docx==1.2.0
PyPDF2==3.0.1

# Pydantic - Version varies by platform
# Windows: 1.10.13 (pre-built wheels)
# Linux:   2.5.3 (full version with Rust)

# Report Generation
reportlab==4.1.0

# Monitoring & Logging
django-debug-toolbar==4.2.0
sentry-sdk==2.18.0
django-structlog==7.1.0
structlog==24.1.0
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.4.1
factory-boy==3.3.0
faker==22.6.0
hypothesis==6.102.3

# Code Quality - Django 6.0 Compatible
black==24.1.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3
mypy==1.8.0
django-stubs==5.2.9

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
arrow==1.3.0
pytz==2024.1

# Optional: API Documentation
drf-spectacular==0.26.5
drf-yasg==1.21.8

# Optional: S3 Storage (for Cloude)
# boto3==1.34.18
# django-storages==1.14.2

# Optional: Monitoring
# prometheus-client==0.19.0
# django-prometheus==2.3.1

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
# uwsgi==2.0.23  # UNIX/Linux only - included in requirements-linux.txt

# Security
cryptography==41.0.7
django-ratelimit==4.1.0
